// src/app/models/model/barcode/barcodeModel_A.ts
export class BarcodeModel_A  
{
    itemArea: string;
    refArea: string;
    lotArea: string;
    qtyArea: number;
    addressArea: string;
    boxArea: string;
    dateArea: Date;
    topBarcodeJSON: string;
    productBarcodeJSON: string;
    barcode:string
    itemCode :string
}

export class BarcodeModelResponse 
{
    page: string;
   
}